Established unit test to evaluate mapping records
Historically, all DTOs used by the `ITopicMappingService` have been `class`es. This unit test establishes a DTO based on the new C# 9.0 `record` type. Initially, this fails. The goal of this branch is to resolve the issues with the underlying services—and, notably, the `TypeMemberInfoCollection`—that currently prevent records from being mapped.