Handle duplicate members in `MemberInfoCollection<T>`
Historically, the `MemberInfoCollection<T>` couldn't support duplicate members. This would occur if there was an overload or a override of a member.

This typically isn't usually a major issue with mapping view models, since they rarely have overloads or overrides. But it's an issue with C# 9.0's new `record` type, since the compiler dynamically generates an override of the `Equals()` operator for records.

A quick fix for this is to operate on a first-come-first-serve basis. If there's a duplicate key, that member will be ignored. As far as the `MemberInfoCollection<T>` is concerned—and, thus, the `TopicMappingService` which depends upon it!—only the first of the overloads or the most derived of the overrides will count.

There might be some issues with overloads where this assumption doesn't make sense—but to address those, we'd need a much more involved workaround, such as an attribute for flagging the priority overrload. That's not an immediate need. In the meanwhile, this satisfies most scenarios, avoids an exception when there's an overload or override, and allows us to map C# 9.0 records.