Introduced the new `TopicReferences` table
The `TopicReferences` table is functionally identical to the existing `Relationships` table, except that it enforces a 1:1 relationship, instead of an 1:n relationship. Notably, with relationships, a topic can have multiple topics related via a relationship key, whereas with reference, a topic can only have a single topic related via a reference key.

(Note: A topic can still have multiple references—just only one reference per key. In that way, the 1:1 vs 1:n terminology gets a bit confusing. If each reference were modeled as a hard-coded column, instead of as a row with an arbitrary `ReferenceKey`, the relationship would be clearer.)