Removed unnecessary prefixes
The `GetTopics` stored procedure had a number of redundant prefixes. This is a stylistic preference, and there's an argument for keeping them—it's more explicit, and it ensures variables are lined up. That said, with longer column names, it also breaks up the value alignment. Generally, in other plaes, we only use them when necessary, so I'm maintaining that here.