Establish a `TopicReferences` user-defined table types
By establishing a custom `TopicReferences` user-defined table types—and corresponding `TopicReferencesDataTable` in C#—we'll be able to pass key/value pairs for topic references to stored procedures, similar to how we pass `AttributeValues` and `TopicList`s today. This is more elegant and less error-prone than e.g. passing a delimited string and attempting to parse it in SQL.