Removed defaults for required parameters
The `UpdateTopic` and `UpdateRelationships` stored procedures both had a default value for `@TopicID` of `-1`, with the latter also having a default value for `@RelationshipKey` of `related`.

These should be considered required fields. There is never expected to be a (legitimate) `TopicID` of `-1` in the database; attempting to execute this would yield an exception. And there's no reasonable need for a default relationship—that should be defined by the content type descriptor.

In practice, we always pass these values; by not including defaults, we better communicate that these are strictly required.