Pass existing `unresolvedTopics` cache when resolving topics
When picking up unresolved relationships in the `SqlTopicRepository.Save()` method, `isRecursive` is always set to `false` and, therefore, the `unresolvedTopics` cache doesn't come into play, as there's no subsequent opportunity for any unresolved relationships to be recovered (since they definitely exist outside the scope of the current operation).

As the private `Save()` overload requires a unresolved relationships cache, however, we were previously constructing a new one for each call. This is a cheap operation. But as we already have one, and it doesn't matter what it contains, we might as well reuse it.

This is a minor optimization.