Introduced new `UpdateReferences` stored procedure
Technically, since this is a 1:n relationship, and not an n:n relationship, this logic could be built into e.g. the `UpdateTopic` stored procedure. To maintain consistency with `UpdateRelationships`, however, and to allow it to be tested independently, I'm setting it up as its own stored procedure.

That said, given that there will only be one of these per update, we'll be adding a `@References` parameter to the `CreateTopic` and `UpdateTopic` stored procedures later, so they don't need to call this separately.