Incorporated `UpdateReferences` into `CreateTopic` and `UpdateTopic`
Incorporated the new `UpdateReferences` stored procedure (1a62c23) into the `CreateTopic` and `UpdateTopic` stored procedures by introducing a new `@References` parameter and relaying it to the `UpdateReferences` stored procedure in the body, alongside the `@TopicID`.

In addition, for the `UpdateTopic` stored procedure, I exposed a new, optional `@DeleteUnmatched` parameter, which maps to the optional `@DeleteUnmatched` parameter on the `UpdateReferences` stored procedure. This allows callers of `UpdateTopic` to delete any unmatched references, which is useful if the `@References` collection is expected to be completed, as opposed to just being a difference.

Finally, the `SqlTopicRepository` has been updated to pass these parameters along in order to ensure these calls remain valid. That said, the `Save()` method is not yet injecting data into this. Further, we may update this later to be handled as a secondary call—similar to  `PersistRelationships()`—just because it works better with `areReferencesResolved`.

Regardless, we'll maintain the _ability_ for the stored procedures to insert references, even if we don't take advantage of it via `SqlTopicRepository.Save()`.