Established `TopicReferenceDictionary` for tracking topic references
A dictionary makes the most sense here since there will be a 1:1 relationship between a `referenceKey` and a `Topic`—but the `referenceKey` can't be a property of the `Topic` since the same `Topic` may be referenced from multiple other `Topic`s.

Unfortunately, none of the out-of-the-box implementations of `IDictionary<TKey, TValue>` allow write operations to be intercepted in order to e.g. track `IsDirty` or handle incoming relationships, as we were able to do with `KeyedCollection`. In fact, even if we wanted to intercept these at each possible entry point, we can't becomes none of the members are marked as `virtual`—and not just on `Dictionary<TKey, TValue>`, but on all other out-of-the-box implementations.

To mitigate this, I've created a `TopicReferenceDictionary` façade which implements `IDictionary<TKey, TValue>`, and then relays those calls to a private `Dictionary<TKey, TValue>` backing field. This allows us to intercept these calls without needing to recreate the under-the-hood logic of a dictionary.

This initial implementation is barebones and doesn't yet include full support for `IsDirty()` tracking, reciprocol relationships, or optimized methods for setting and retrieving topics by key. Those will come in subsequent commits. For now, this establishes the basic framework for the façade and, importantly, frames out the interception points we'll need for e.g. `Add()`, `Remove()`, and `Clear()`.