Prevent a topic from referencing itself
This should generally be a valid assumption, and prevents obvious scenarios that will lead to circular loops. With existing topic references—such as `DerivedTopic`—this is the main piece of business logic we need, and thus staves off the need to handle business logic via reflection, as we need to do with the `AttributeValueCollection`.

(I imagine we'll eventually want to handle business logic in a similar way, but it's a fairly complex addition, and so not something I want to add until it's necessary.)