Incorporate `IncomingRelationships` into `TopicReferenceDictionary`
We stored relationships in both `Topic.Relationships` as well as `Topic.IncomingRelationships` in order to maintain a reciprocal relationship.

We'll want to do something similar with `Topic.References` via the `TopicReferenceDictionary`. If a `Topic` is targeted by multiple references, however, that reciprocol relationship can't be modeled via a `TopicReferenceDictionary` since it models a 1:1 relationship per `referenceKey`.

That said, we can treat the incoming references as a relationship and store those in the same `Topic.IncomingRelationships` property used for normal relationships, which _does_ support a 1:n relationship per key. And since `IncomingRelationships` is only maintained for convenience, and not persisted, this doesn't interfere with the backend data modeling—unless, of course, there is a relationship and a reference with the same key. Semantically, that probably doesn't make much sense usually, and is something we should avoid regardless—but if it happens, it will work, but result in an ambiguous `IncomingRelationships` collection for that key.

(It's worth noting that this is already an potential design concern with `IncomingRelationships` since two `ContentTypeDescriptors` can have the same `RelationshipKey`, even if they refer to potentially distinct concepts. Designer beware!)