Introduce `SetTopic()` method on `TopicReferenceDictionary`
The `SetTopic()` will `Add()` a new topic, set an existing topic, or remove a `null` topic reference. It also allows `IsDirty` to be disabled as part of the write, which is important for establishing references on existing entities (e.g., as part of `ITopicRepository.Load()`).

Callers can still use e.g. `Add()` or `this[]`, obviously, but I expect `SetTopic()` to be the most convenient entry point.