Persist topic references as part of `SqlTopicRepository.Save()`
The logic for this was scaffolded in a previous commit (8a7d5d1), but the data couldn't be populated until we had established `Topic.References` to pull from. Now that that is complete (573a800), we're able to close that loop.

Note: We'll likely move this to be treated as an ancillary call, similar to `PersistRelationships()`, in a later commit. For now, however, this should function.