Account for `References` in `unresolvedRelationships`
As with relationships, if any references point to unresolved topics—i.e., topics that haven't yet been saved—then we want to attempt to save them again later, once a recursive save has completed.

Unlike relationships, this will result in the `References` being saved twice, since they're persisted as part of the call to the `UpdateTopic` or `CreateTopic` stored procedures. We'll likely change that in a subsequent commit.