Established `PersistReferences()` helper method
While the `CreateTopic` and `UpdateTopic` stored procedures accept a `@TopicReferences` parameter, it makes more sense to save topic references as a separate query so that we don't need to do a second save in the case that there are unresolved references. This logic works very similar to the existing `PersisRelations()` method.