Conditionally exclude references, if not supplied
If `@DeleteUnmatched` is true (`1`) and `@References` is null or empty, then `UpdateReferences` will delete all references. As we conditionally remove references, this potentially causes a problem.

By putting in this condition we introduce other potential issues since we now can't delete all references. I need to think through this some more. In the meanwhile, this prevents accidental deletion of all references.