Refactor `isDirty` handling to bypass attributes if `!areAttributesDirty`
We previously defined `areAttributesDirty`, and factored it into the calculation of `isDirty`. Now, the check for mismatched extended attributes is moved from `isDirty` to instead factor into `areAttributesDirty`. That doesn't change the logic of `isDirty`, since it still relies on `areAttributesDirty`. But it allows us to conditionally exclude the `@Attributes` and `@ExtendedAttributes` from `UpdateTopics` if no attributes have been modified, and there aren't any mismatches to persist.

As part of this, `DeleteUnmatched` is reverted back to `false`, as we don't want to delete any unmatched `@Attributes`—nevertheless `@References`!—if the collection isn't passed. (Unfortunately, SQL doesn't provide a way to differentiate between a null and empty user-defined, table-valued type.)