Conditionally exclude `Key` and `ContentType` unless they've changed
In a previous commit, I updated `UpdateTopic` so that it would only save `@Key` or `@ContentType` if they were provided, thus maintaining consistency with the rest of `UpdateTopic`, which allows omitting values if they haven't changed (a74389d).

In this update, I've made `SqlTopicRepository` aware of this change by conditionally excluding the `@Key` and `@ContentType` parameters unless `Topic.IsDirty()` or `Topic.IsNew`. (If a topic is new, the core attributes should _always_ be marked as dirty, so the logic is a bit redundant, but it helps reinforce that the logic.)