Don't execute `UpdateTopic` if neither the topics nor attributes are dirty
Currently, we skip `Save()` entirely if `isDirty` isn't true. And we now exclude `@Attributes` and `@ExtendedAttributes` if `areAttributesDirty` is false, and `@Key` and `ContentType` if `isTopicDirty` is false.

But if both `isTopicDirty` and `areAttributesDirty` are false, but `isDirty` isn't—e.g., due to `areRelationshipsDirty` or `areReferencesDirty`—then there's no need to execute `UpdateTopic` at all. This update bypasses the `ExecuteNonQuery()` in this scenario.

This should offer a marginal performance improvement when dealing with recursive saves of a large topic graph, as is often the case when importing topics.