Don't construct `attributeValues`, `extendedAttributes` if `!areAttributesDirty`
If `areAttributesDirty` is false, then neither `@Attributes` nor `@ExtendedAttributes` will be sent to the `UpdateTopic` stored procedure, so there's no purpose to populating the `attributeValues` or `extendedAttributes` collections.

As part of this, I consolidated the population of `attributeValues` so that it was broken up by the population of `extendedAttributes`. (There are two parts to populating `attributeValues`: adding the attributes that have changed, then adding the attributes that have either not changed or are missing.)

This provides a very marginal performance benefit by excluding the population of unused collections. The collections are still initialized, however, since conditionally excluding them introduces additional complexities with null checking down the road.