Updated unit tests for `DerivedTopic`
The previous unit tests for `DerivedTopic` evaluated the `Attributes` as a backing store to ensure the value was properly reflected there. As `DerivedTopic` is now stored in `Topic.References` instead of `Topic.Attributes` (0ec079c), these tests needed to be updated.

Note: These tests may not be as meaningful at this point since we'd expect these to reference the same object, instead of requiring a serialization of the value. We may want to revisit if these provide value later.