Allow topic references to inherit from derived topics
This maintains familiar functionality from `topic.Attributes.GetValue()`, and ensures that any topic references which previously relied upon inheritance when they were stored as attributes continue to work now that they're stored as `TopicReferences`.

(This excludes calls to `Topic.DerivedTopic`, of course, as that would introduce an infinite loop!)

This includes three new unit tests to ensure that inheritance works as expected—and can be optionally disabled.