Introduced support for `TopicReferences` in `TopicMappingService`
The `TopicMappingService` already has support for topic references—but it previously assumed that they would be stored as attributes ending with `Id`. That support is maintained (for now) for backward compatibility. In addition, however, the `SetPropertyAsync()` method now _also_ checks the `Topic.References` collection for an exact match and, if found, prefers that over the attribute-based topic reference.