Introduced support for `TopicReferences` in `ReverseTopicMappingService`
There was already support for `ModelType.Reference`, but it was based on storing the referenced topic's `Id` as an attribute. As part of that, it enforced that the attribute name must, by convention, end with an `Id`. That validation is now removed, and the topic reference is now stored in `Topic.References` instead of `Topic.Attributes`.

As with the corresponding `TopicMappingService` implementation (e5b87eb), the `ReverseTopicService` implementation maintains backward compatibility with legacy attribute-based topic references. It does so by relying on the legacy `{ReferenceKey}Id` convention: If the `AttributeKey` ends with `Id`, then it will be stored in `Topic.Attributes`; otherwise, it will be stored in `Topic.References`.

Note that the `Id` check is case-sensitive. We don't expect the `ReferenceKey` used in `Topic.References` to end in `Id`—and, in fact, remove the `Id` as part of the migration script (db35570). But it is possible they they will end in an `id` (e.g., `Aid` or `Android`). By keeping this case sensitive, we help avoid false positives.