Updated unit tests to use `DerivedTopic` from `Topic.References`
This included removing the explicit `[AttributeKey()]` setting it to use `TopicId` (instead of the default of `DerivedTopic`), changing the `AttributeDescriptor` to be named `DerivedTopic` instead of `TopicId`, and, finally, removing the explicit setting of the `DerivedTopic` reference in the unit test based on the `Topic.Attributes` value.

While I was at it, I extended the unit test to confirm that the `DerivedTopic` has the correct key.

Note: This does _not_ test the backward compatible support for `{ReferenceKey}Id` names that _should_ be stored in `Topic.Attributes`. I may reintroduce a unit test for that again in the future. For now, this isn't an immediate priority, and especially as this functionality will likely be removed in a future release—possibly even before the next release.