Removed `TopicReferences` support from `UpdateTopic`
Previously, the `UpdateTopic` stored procedure had been updated to accept an "optional" `@References` parameter which, if present, would be relayed to `UpdateReferences` (8a7d5d1).

Unfortunately, this introduces confusion as there's no way to differentiate between a `null` or empty table-valued typed parameter. As such, this leads to one of two scenarios: a) if processing is skipped if `@References` is empty, then there's no way to delete all topic references, but alternatively b) if no `@References` parameter is defined, but `@DeleteUnmatched` is, then all topic references would be deleted, assuming we didn't check for empty.

Ideally, we'd be able to differentiate between null and empty here, and there wouldn't be a problem. In absence of that, however, we also don't want to force callers to include all topic references in order to use `@DeleteUnmatched`, assuming they only want to update e.g. the attributes. This is especially true since, in `SqlTopicRepository.Save()`, we don't pass `@References` since it may contain unresolved references; in those cases, the save is deferred.

To mitigate that, the `@References` parameter has been removed. This means updates to `@References` must be made via the `UpdateReferences` stored procedure.

Note: These conflicts don't exist on the `CreateTopic` stored procedure, since `@DeleteUnmatched` has no relevance assuming the topic is new. As such, I'm maintaining the `@References` parameter on `CreateTopic`, thus giving it a bit more flexibility. The `SqlTopicRepository.Save()` won't use this, but it may be useful for other calls, and especially utilities or maintenance scripts.

That said, I'm a bit conflicted by the (further) inconsistency between the `CreateTopic` and `UpdateTopic` signatures, and may revisit this later, yet again.