Introduced `UpdateAttributes` stored procedure
Moves the logic from `UpdateTopic` into a standalone stored procedure. This is useful since the logic for this is becoming fairly complex. It also allow the `Attributes` table to be easily updated independent of other tables. (Technically, the `UpdateTopic` stored procedure also allowed this, but this makes it more explicit and intuitive.)