Convert `References.IsDirty` from property to method
This is more consistent with other collections—such as `AttributeValueCollection` and `RelatedTopicCollection`—as well as `Topic` itself. Technically, this is just operating as a bit, and could continue to be a property, but treating it as a method gives us more flexibility in the future to e.g., add optional parameters, if needed, or track individual references, as we do with attributes.