Added a `MarkClean()` method to `RelatedTopicCollection`
There was previously no way to mark all relationships as clean without looping through the list of `NamedTopicCollections` and flipping the `IsDirty` value of each. That's all the new `MarkClean()` method does as well, but it centralizes the logic in the `RelatedTopicCollection`, and offers consistency with other collections that use the `IsDirty()`/`MarkClean()` convention (e.g., `AttributeValueCollection`).