Changed return type for `ISqlTopicRepository.Save()` to `void`
Previously, the `ISqlTopicRepository.Save()` method—and, thus, all of its implementations—expected a return type of `int`, representing the `Topic.Id`. This is most valuable when saving a new topic, which wouldn't have had an `Id` previously assigned. In most cases, however, this is never expected or needed. And, in the few cases where it is, it can easily be retrieved from `Topic.Id` instead—which is always available, since it's a required parameter of `Save()`.