Established a reusable `TopicPropertyDispatcher` class
When an `AttributeValue` is added to the `AttributeValueCollection`, it checks to see if a corresponding property on the associated `Topic` exists based on the `AttributeKey` (which is expected to match the property name) and the presence of an `[AttributeSetter]` attribute. If it does, the value is set using that property, instead of being added directly to the collection. This prevents the collection from bypassing the business logic of a property setter, which could prevent data validation from occurring, or interfere with internal state tracking of the `Topic` instance.

The new (`internal`) `TopicPropertyDispatcher` class extracts the code for accomplishing this, and places it it in a reusable utility that can be applied to other collections, thus preventing this from being used _exclusively_ by `AttributeValueCollection`.

This has the additional benefit of allowing us to simplify the code for the `AttributeValueCollection`, while moving the logic for enforcing business to a class that's focused entirely on that task. While I was at it, I greatly improved the documentation for this functionality, so it's hopefully a bit clearer. (The abstraction makes this already complicated logic even more difficult to reason through, but hopefully the added remarks aid in that.)