Handle `TargetInvocationException` to deregister dispatch
If a `TargetInvocationException` is thrown—for any reason—while enforcing business logic, we want to ensure that registration of the current property is removed from the `TopicPropertyDispatch`, if it hasn't been already. Otherwise, if we don't, then a subsequent attempt to set the same attribute on the topic will bypass the business logic, even if it's a different value.

To prevent this loophole, we explicitly catch the `TargetInvocationException`, check to see if the key is still in the `PropertyCache`, and, if it is, remove it. (We have to check because, hypothetically, a property could throw an exception _after_ it's successfully added the item to the collection, in which case `Enforce()` would have been already called a second time, and the request will no longer be in the `PropertyCache`.)

Afterwards, we rethrow the exception so that callers are aware that the action (potentially) failed.