Rethrow _inner_ exception on `TopicPropertyDispatcher.Enforce()`
When calling `SetPropertyValue()`, the most likely exception is a `TargetInvocationException`, which will occur if _any_ error occurs while invoking the corresponding property. That effectively buries the actual reason, and the `TargetInvocationException` really only makes sense if you're aware of the internal workings of OnTopic.

To avoid that confusion, and provide a more meaningful exception, we're relying on `ExceptionDispatchInfo` to rethrow the `InnerException`, thus ensuring that the actual error bubbles up, instead of the `TargetInvocationException`.

As part of this, the unit tests which look for the `TargetInvocationException` need to be updated to expect the actual underlying exception. (This also makes those unit tests more meaningful, as they're not tripping on any error, but the specific error that the target property is expected to throw.)