Move specific collections to more relevant namespaces
As with the underlying .NET CLR, collections in the general `Collections` namespaces should be reserved for general-purpose collection. By contrast, collections that are specific to a particular purpose should be distributed among their relevant namespaces. So, for example, `AttributeValueCollection` should be in `OnTopic.Attributes`, alongside the `AttributeValue` type itself.

There wasn't a relevant namespace for the relationship-related collections, which includes `RelatedTopicCollection`, `NamedTopicCollection`, and `TopicReferenceDictionary`. Placing them in `OnTopic.Relationships` introduces a conflict with the existing `Relationships` enum used by the mapping service. As such, I chose `OnTopic.References`. (It doesn't strictly matter, as this is being used for both relationships—i.e., 1:n mappings—and topic references—i.e., 1:1 mappings. Technically, they could each be described as either relationships or references.)

This also covers the `OnTopic.Internal.Collections` namespace, with the `MappedTopicCache` being moved to `OnTopic.Internal.Mapping` (alongside the `MappedTopicCacheEntry`).

As part of this, the namespaces needed to be updated in a number of other classes to ensure the new class locations could be resolved—even if that's only needed for the XML Docs.