Removed warning suppressions for injection vulnerabilities
In `Microsoft.CodeAnalysis.NetAnalyzers` 5.0.0 and 5.0.1, there was a bug that caused multiple exceptions when assessing various injection vulnerabilities (CA3001, 3003, 3006, 3008, 3009, 3011, 3012) (see dotnet/roslyn-analyzers#4495). This was resolved in 5.0.3, so the warnings no longer need to be suppressed.

Unfortunately, a couple of other suppressions that I had understood to be resolved still appear to be needed, so that will require additional evaluation to ensure there isn't some confounding variable.

Note: Normally these would be suppressed in the `GlobalSuppressions` class or via an inline `pragma` directive. Perhaps because these analyzers were throwing exceptions, that didn't satisfy code analysis, and thus they had to be handled globally as the `<NoWarn />` element in the `csproj`.