Moved `TypeCollection` to `OnTopic.Lookup`, marked `internal`
While the `TypeCollection` is conceivably useful for broader purposes, it is currently exclusively used by the `StaticTypeLookupService` as a private backing field. As such, there's no need for it to be marked as `public`, so I've marked it as `internal`.

Further, as an `internal` dependency, there's no need to it to be explicitly organized under `OnTopic.Internal` (which is meant to distinguish `public` classes that aren't intended for public use). As such, it can be moved to `OnTopic.Lookup`, next to the `StaticTypeLookupService.

If other internal classes want to use it, `OnTopic.Collections` will make more sense, since it's a fairly general collection. As an `internal` dependency, we can move it at a later date if, in needed.