Rename `TypeMemberInfoCollection` to `MemberDispatcher`
The `TypeMemberInfoCollection` is a dynamic dispatch service supporting late-binding of members and properties—but one would never guess that from the class name.

To remedy that, the, the class is being renamed to the more descriptive `MemberDispatcher`, similar to the (new) `TopicPropertyDispatcher`.

As these are (now) `internal`, this shouldn't have any (further) impact on any external libraries.