Extracts `TypeMemberInfoCollection` from `MemberDispatcher`
Initially, the `MemberDispatcher` was a keyed collection of `MemberInfoCollection` instanced, keyed by `Type`. This grew into a dynamic dispatch service, and thus the rename to `MemberDispatcher` (50019c7). As the base class renamed a `KeyedCollection<>`, however, it has a lot of members that are exclusively needed for internal access, and only act to clutter—and confuse—the class's interface.

To mitigate this, the underlying collection has been extracted, and moved into a reestablished `TypeMemberInfoCollection` which exclusively contains the collection-related logic. This is then used as a private cache in the `MemberDispatch` instead of as a base class. This dramatically cleans up the interface of the `MemberDispatch`, and better distinguishes between responsibilities.