Moved non-dispatch related methods to `TypeMemberInfoCollection`
Some methods included in `MemberDispatcher` aren't specific to the dynamic dispatch functionality, per se, and have broader relevance to the underlying `TypeMemberInfoCollection` used as a cache. As such, the implementations of these have been moved to `TypeMemberInfoCollection`.

Unfortunately, these continue to be called from internal implementors. It's unclear if all of those are required. But while those are evaluated and confirmed, I am maintaining the original signatures as internal passthroughs to the `TypeMemberInfoCollection`.