Removed unused passthroughs from `MemberDispatcher`
In the previous commit, a number of methods that are more relevant to `TypeMemberInfoCollection` were moved to there—but passthroughs were maintained on `MemberDispatch`. In this update, the passthroughs that were not used—notably, those which didn't accept generic type parameters—were removed, so as to avoid confusion.

I'd stil like to reevaluate the ones that remain at some point, but those can be revisited later, if necessary.