Enable entry points to map disabled topics
It should be up to callers as to whether or not a topic that `IsDisabled` should be mapped. Generally, it is expected that it won't be—and, thus, we enforce this via e.g. the `[ValidateTopic]` action filter attribute. But, there are other scenarios where we wish this to be possible, such as the OnTopic Editor, as otherwise disabled topics can't be edited.