Prevent disabled topics from being mapped as relationships, references
While callers may request that an `IsDisabled` topic be mapped (fde9c83), they have no way of validating referenced topics within the topic graph. We should therefore assume that child topics that are disabled should not be mapped, to prevent potential pollution of the topic graph. This will still allow disabled topics to be e.g. edited in OnTopic Editor, but will prevent them from showing up as children, relationships, or references off topics.

In the future, we may wish to provide a way of configuring this per request, possibly via an options object or even a validator delegate. But that would break the interface and adds a bit of complexity. For now, we expect the idea of enabling disabled root topics to be mapped, but not disabled referenced topics, should satisfy most requirements.