Ensures that `Rollback()` correctly sets `AttributeValue.IsDirty`
By default, when a new version is loaded, all attributes will be set to `!IsDirty`. The `Rollback()` method is intended to check if any attributes are new or have changed and, if so, set them to `IsDirty` to ensure that they are changed. Previously, it was setting `IsDirty` to `false`—which is what the value already would have been.