Remove setting of core attributes on `Rollback()`
Previously, core attributes such as `Key`, `ContentType`, and `ParentId` were stored as attributes, where they could, conceivably, be versioned. This introduced a lot of complexity, and potential for error. (For instance, if a rollback introduces a duplicate `Key`, or attempts to move a topic back to a previous `ParentId`.) As such, we moved these to the `Topics` table, where they are not versioned. As a result, we don't expect `Rollback()` to chenge these values. Furthermore, there's no need to manage their `IsDirty` state.